LPS-98190 move MetatypePlugin back to ant-bnd to support maven as well
