+++ modules/sdk/ant-bnd/src/main/java/com/liferay/ant/bnd/metatype/MetatypePlugin.java
@@ -17,49 +17,50 @@
  * under the License.
  */
 
-package com.liferay.gradle.plugins.workspace;
-
-import java.util.Collection;
+package com.liferay.ant.bnd.metatype;
 
 import aQute.bnd.header.Parameters;
 import aQute.bnd.make.metatype.MetaTypeReader;
 import aQute.bnd.osgi.Analyzer;
 import aQute.bnd.osgi.Clazz;
-import aQute.bnd.osgi.Clazz.QUERY;
 import aQute.bnd.osgi.Constants;
 import aQute.bnd.osgi.Jar;
 import aQute.bnd.service.AnalyzerPlugin;
 
- /**
+import java.util.Collection;
+
+/**
  * This class is responsible for meta type types. It is a plugin that can
  *
  * @author aqute
  */
 public class MetatypePlugin implements AnalyzerPlugin {
 
- 	public boolean analyzeJar(Analyzer analyzer) throws Exception {
+	public boolean analyzeJar(Analyzer analyzer) throws Exception {
+		Parameters parameters = analyzer.parseHeader(
+			analyzer.getProperty(Constants.METATYPE));
 
- 		Parameters map = analyzer.parseHeader(analyzer.getProperty(Constants.METATYPE));
+		Jar jar = analyzer.getJar();
 
- 		Jar jar = analyzer.getJar();
-		for (String name : map.keySet()) {
-			Collection<Clazz> metatypes = analyzer.getClasses("", QUERY.ANNOTATED.toString(),
-					"aQute.bnd.annotation.metatype.Meta$OCD", //
-					QUERY.NAMED.toString(), name //
-			);
-			for (Clazz c : metatypes) {
-				analyzer.warning(
-						"%s annotation used in class %s. Bnd metatype annotations are deprecated as of Bnd 3.2 and support will be removed in Bnd 4.0. Please change to use OSGi Metatype annotations.",
-						"aQute.bnd.annotation.metatype.Meta$OCD", c);
-				jar.putResource("OSGI-INF/metatype/" + c.getFQN() + ".xml", new MetaTypeReader(c, analyzer));
+		for (String name : parameters.keySet()) {
+			Collection<Clazz> metatypeClasses = analyzer.getClasses(
+				"", Clazz.QUERY.ANNOTATED.toString(),
+				"aQute.bnd.annotation.metatype.Meta$OCD",
+				Clazz.QUERY.NAMED.toString(), name);
+
+			for (Clazz c : metatypeClasses) {
+				jar.putResource(
+					"OSGI-INF/metatype/" + c.getFQN() + ".xml",
+					new MetaTypeReader(c, analyzer));
 			}
 		}
+
 		return false;
 	}
 
- 	@Override
+	@Override
 	public String toString() {
 		return "MetatypePlugin";
 	}
 
- }
+}
