+++ modules/sdk/gradle-plugins-workspace/build.gradle
@@ -51,6 +51,7 @@ copyRepo {
 
 dependencies {
 	compile group: "com.bmuschko", name: "gradle-docker-plugin", version: "3.6.2"
+	compile group: "com.liferay", name: "com.liferay.ant.bnd", version: "3.1.0"
 	compile(group: "com.liferay", name: "com.liferay.gradle.plugins", version: "4.0.66") {
 		exclude group: "biz.aQute.bnd", module: "biz.aQute.bnd"
 	}
