+++ source-formatter.properties
@@ -17,7 +17,6 @@
         **/tools/sdk/**,\
         **/webapps/**,\
         definitions/web-facesconfig_1_2.xml,\
-        modules/sdk/gradle-plugins-workspace/src/main/java/com/liferay/gradle/plugins/workspace/MetatypePlugin.java,\
         modules/yarn-1.13.0.js,\
         portal-impl/src/portal-ext.properties,\
         portal-web/docroot/dtd/web-facesconfig_1_2.xml
